Fix critical and high-severity bugs across serial, transport, and UI
  Backend:
  - Remove unused, unconfigured updater plugin that panicked every native
    build at launch (registration, dependency, and capability grant)
  - Reject embedded newlines in g_command so the allow-list (incl. the M997
    firmware-flash block) can't be bypassed with a smuggled second line
  - Store temperatures as i16 instead of u8 so readings >=256C and faulted
    thermistors no longer wrap to 0C; accept negatives in the M105 regex
  - Complete serial reads on Marlin's ok/error ack with an idle fallback and
    a 30s ceiling, replacing the 300ms cap that cut off G28/M109/G29
  - Clear the shared serial sender when a worker exits on a fatal error so
    reconnection works after a cable-yank (guarded by a connection generation)

  Frontend:
  - Emit relative extruder moves (M83 + G1 E) and fix axis command formatting
    in moveAxis instead of building corrupt absolute E targets
  - Reflect the disconnect when switching transports so the UI can't stay
    stuck "connected" to nothing
  - Stop M27 C from clobbering an active print to IDLE; drive state from M27
  - Surface a disconnect when replacing a live WebSocket, add an upload
    inactivity watchdog, and clear the connector spinner on close
  - Mount tab panels lazily so only the active tab's scene and pollers run
  - Ignore duplicate/concurrent USB connects and keep a rejected attempt from
    tearing down the live connection's listeners

+++ src-tauri/src/printer_core/structs.rs
@@ -75,18 +75,22 @@ pub struct AxePositions {
 }
 
 /// M105 - Get Extruder Temperature
+///
+/// `i16` rather than `u8`: hotends routinely exceed 255 °C (ABS/PA/PC), which
+/// would wrap a `u8` to 0 and report "cold" while the heater is at full power;
+/// a faulted thermistor also reports negative values.
 #[derive(Debug, Serialize, Deserialize, Default)]
 pub struct Temperatures {
-    pub bed: u8,
-    pub bed_set: u8,
-    pub e0: u8,
-    pub e0_set: u8,
-    pub e1: u8,
-    pub e1_set: u8,
-    pub e2: u8,
-    pub e2_set: u8,
-    pub e3: u8,
-    pub e3_set: u8,
+    pub bed: i16,
+    pub bed_set: i16,
+    pub e0: i16,
+    pub e0_set: i16,
+    pub e1: i16,
+    pub e1_set: i16,
+    pub e2: i16,
+    pub e2_set: i16,
+    pub e3: i16,
+    pub e3_set: i16,
 }
 
 /// Com configuration
